review comments + clippy

+++ tests/tests/test_mev_commission.rs
@@ -63,7 +63,7 @@ async fn test_mev_commission() {
     assert!(account.history.arr[0].epoch == 0);
     assert!(account.history.arr[0].mev_commission == 42);
     assert!(account.history.arr[0].mev_earned == 12324); // fixed point representation
-    assert!((account.history.arr[0].mev_earned as f64 / 100.0) == 123.24 as f64);
+    assert!((account.history.arr[0].mev_earned as f64 / 100.0) == 123.24_f64);
     // TODO this is causing a hash mismatch issue
     // fixture.advance_num_epochs(1).await;
